# api/app/schemas.py
"""The HTTP API contract: request and response models (D10). The frontend's types are
generated from these through FastAPI's OpenAPI schema.

Response models extend the tables' Base classes (app/models/), so shared fields are declared
once. Those fields have defaults for the tables, so each such model sets
json_schema_serialization_defaults_required: every field stays required in the OpenAPI
schema, and TypeScript gets `number | null` rather than an optional field.
"""

import uuid
from typing import Literal

from sqlmodel import Field, SQLModel

from app.core.config import MAX_UPLOAD_BYTES
from app.models import Analysis, AnalysisBase, DetectionBase, Video, VideoBase

AnalysisState = Literal["awaiting_upload", "queued", "processing", "done", "failed"]


class AnalysisPublic(AnalysisBase):
    model_config = {"json_schema_serialization_defaults_required": True}
    state: AnalysisState  # derived from the video's and the analysis's status


class VideoPublic(VideoBase):
    model_config = {"json_schema_serialization_defaults_required": True}
    id: uuid.UUID
    analysis: AnalysisPublic

    @classmethod
    def from_db(cls, video: Video, analysis: Analysis | None) -> "VideoPublic":
        """The server derives the UI's state: awaiting the upload, then the analysis's status,
        or failed with the reason when the upload was rejected (D9)."""
        if analysis is not None:
            public = AnalysisPublic.model_validate(analysis, update={"state": analysis.status})
        elif video.status == "rejected":
            public = AnalysisPublic(state="failed", error=video.error)
        else:
            public = AnalysisPublic(state="awaiting_upload")
        return cls.model_validate(video, update={"analysis": public})


class VideoCreate(SQLModel):
    id: uuid.UUID  # generated by the browser, so a retried request cannot create a second video
    original_filename: str = Field(min_length=1, max_length=255)
    declared_size: int = Field(ge=1, le=MAX_UPLOAD_BYTES)


class UploadTarget(SQLModel):
    method: Literal["PUT"]
    url: str
    headers: dict[str, str]  # must be sent exactly as given: they are part of the signature


class VideoCreated(SQLModel):
    id: uuid.UUID
    upload: UploadTarget


class DetectionPublic(DetectionBase):
    model_config = {"json_schema_serialization_defaults_required": True}


class DetectionsPublic(SQLModel):
    sample_fps: float
    display_threshold: float  # the UI shows only detections at or above this confidence
    detections: list[DetectionPublic]


class PlaybackPublic(SQLModel):
    url: str  # signed GET URL for the uploaded file


class Health(SQLModel):
    status: Literal["ok"]