# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hot-dog detector API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthz",
        "operationId": "healthz_api_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Create Video",
        "description": "Register a video and sign its upload. Idempotent: the browser generates the id, so a\nretried request returns the same upload instead of creating a second video. New videos are\ncapped per hour and per day (UPLOAD_LIMITS); a retry is never refused.",
        "operationId": "create_video_api_videos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Read Video",
        "operationId": "read_video_api_videos__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}/playback": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Read Playback",
        "operationId": "read_playback_api_videos__video_id__playback_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}/detections": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Read Detections",
        "operationId": "read_detections_api_videos__video_id__detections_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}/reprocess": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Reprocess Video",
        "description": "Delete the results and run the analysis again (D3a).",
        "operationId": "reprocess_video_api_videos__video_id__reprocess_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}/events": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Follow Video",
        "description": "Send the video now and again on every change (D7). The stream ends after a finished\nanalysis, or after STREAM_LIMIT_S, and the browser reconnects. FastAPI validates and\nserializes each video, documents it in the OpenAPI schema, and adds `: ping` while idle.",
        "operationId": "follow_video_api_videos__video_id__events_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "$ref": "#/components/schemas/VideoPublic"
                      }
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisPublic": {
        "properties": {
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts",
            "default": 0
          },
          "frames_sampled": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frames Sampled"
          },
          "detection_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Count"
          },
          "state": {
            "type": "string",
            "enum": [
              "awaiting_upload",
              "queued",
              "processing",
              "done",
              "failed"
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "progress",
          "error",
          "attempts",
          "frames_sampled",
          "detection_count",
          "state"
        ],
        "title": "AnalysisPublic"
      },
      "DetectionPublic": {
        "properties": {
          "t_seconds": {
            "type": "number",
            "title": "T Seconds"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "x1": {
            "type": "number",
            "title": "X1"
          },
          "y1": {
            "type": "number",
            "title": "Y1"
          },
          "x2": {
            "type": "number",
            "title": "X2"
          },
          "y2": {
            "type": "number",
            "title": "Y2"
          }
        },
        "type": "object",
        "required": [
          "t_seconds",
          "confidence",
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "title": "DetectionPublic"
      },
      "DetectionsPublic": {
        "properties": {
          "sample_fps": {
            "type": "number",
            "title": "Sample Fps"
          },
          "display_threshold": {
            "type": "number",
            "title": "Display Threshold"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/DetectionPublic"
            },
            "type": "array",
            "title": "Detections"
          }
        },
        "type": "object",
        "required": [
          "sample_fps",
          "display_threshold",
          "detections"
        ],
        "title": "DetectionsPublic"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "Health"
      },
      "PlaybackPublic": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "PlaybackPublic"
      },
      "UploadTarget": {
        "properties": {
          "method": {
            "type": "string",
            "const": "PUT",
            "title": "Method"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          }
        },
        "type": "object",
        "required": [
          "method",
          "url",
          "headers"
        ],
        "title": "UploadTarget"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoCreate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Original Filename"
          },
          "declared_size": {
            "type": "integer",
            "maximum": 209715200.0,
            "minimum": 1.0,
            "title": "Declared Size"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "declared_size"
        ],
        "title": "VideoCreate"
      },
      "VideoCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadTarget"
          }
        },
        "type": "object",
        "required": [
          "id",
          "upload"
        ],
        "title": "VideoCreated"
      },
      "VideoPublic": {
        "properties": {
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisPublic"
          }
        },
        "type": "object",
        "required": [
          "original_filename",
          "duration_s",
          "codec",
          "id",
          "analysis"
        ],
        "title": "VideoPublic"
      }
    }
  }
}